+++ lib/inc/layer.h
@@ -52,7 +52,7 @@ public:
      * @param weights Vector of neuron weights-vector
      * @param biases Vector of neuron biases.
      */
-    Layer( const uint& nbr_of_inputs, const vector<Eigen::VectorXf>& weights, const vector<float>& biases );
+    Layer( const uint& nbr_of_inputs, const vector<Eigen::VectorXd>& weights, const vector<double>& biases );
 
     ~Layer();
 
@@ -63,61 +63,61 @@ public:
      * @param x_in Input signal.
      * @return true if successful.
      */
-    bool feedForward( const Eigen::VectorXf& x_in );
+    bool feedForward( const Eigen::VectorXd& x_in );
 
     /**
      * Sets the weights-vector in each neuron of this layer.
      * @param weights Vector of neuron weights-vector.
      * @return true if successful
      */
-    bool setWeights( const vector<Eigen::VectorXf>& weights );
+    bool setWeights( const vector<Eigen::VectorXd>& weights );
 
     /**
      * Sets the weights in this layer.
      * @param weights The weight matrix
      * @return true if successful
      */
-    bool setWeights( const Eigen::MatrixXf& weights );
+    bool setWeights( const Eigen::MatrixXd& weights );
 
     /**
      * Sets the same weight for all neurons and all intputs
      * @param weights Weight value.
      */
-    void setWeight( const float& weight );
+    void setWeight( const double& weight );
 
     /**
      * Returns the current weight matrix. It does not reassemble the matrix
      * ( see updateWeightMatrixAndBiasVector() )
      * @return
      */
-    const Eigen::MatrixXf& getWeigtMatrix() const { return m_weightMatrix; }
+    const Eigen::MatrixXd& getWeigtMatrix() const { return m_weightMatrix; }
 
     /**
      * Sets the bias of each neuron in this layer.
      * @param biases Vector of neuron biases.
      * @return true if successful
      */
-    bool setBiases( const vector<float>& biases );
+    bool setBiases( const vector<double>& biases );
 
     /**
      * Sets the bias of each neuron in this layer.
      * @param biases Vector of neuron biases.
      * @return true if successful
      */
-    bool setBiases( const Eigen::VectorXf& biases );
+    bool setBiases( const Eigen::VectorXd& biases );
 
     /**
      * Sets the same bias for all neurons
      * @param bias Bias value.
      */
-    void setBias( const float& bias );
+    void setBias( const double& bias );
 
     /**
      * Returns the current bias vector. It does not reassemble the vector
      * ( see updateWeightMatrixAndBiasVector() )
      * @return
      */
-    const Eigen::VectorXf& getBiasVector() const { return m_biasVector; }
+    const Eigen::VectorXd& getBiasVector() const { return m_biasVector; }
 
     /**
      * Resets all weights and biases of each neuron in this layer
@@ -130,13 +130,13 @@ public:
      * after executing feedForward().
      * @return Output activation Vector
      */
-    const Eigen::VectorXf& getOutputActivation() const { return m_activation_out; }
+    const Eigen::VectorXd& getOutputActivation() const { return m_activation_out; }
 
     /**
      * Get the input activation of this layer. This is set after calling feedForward().
      * @return Input activation.
      */
-    const Eigen::VectorXf& getInputActivation() const { return m_activation_in; }
+    const Eigen::VectorXd& getInputActivation() const { return m_activation_in; }
 
     /**
      * This is an intermediate result of calling feedForward(). It is the weighted input,
@@ -144,7 +144,7 @@ public:
      * function.
      * @return weighted input.
      */
-    const Eigen::VectorXf& getWeightedInputZ() const { return m_z_weighted_input; }
+    const Eigen::VectorXd& getWeightedInputZ() const { return m_z_weighted_input; }
 
     /**
      * This function computes the backpropagation error in case this is the output layer.
@@ -152,7 +152,7 @@ public:
      * @param expectedNetworkOutput The desired network output.
      * @return Return true if operation was successful. Otherwise false
      */
-    bool computeBackpropagationOutputLayerError( const Eigen::VectorXf& expectedNetworkOutput );
+    bool computeBackpropagationOutputLayerError( const Eigen::VectorXd& expectedNetworkOutput );
 
     /**
      * Computes the backpropagation error in this layer. The backpropagation error can be accessed
@@ -160,7 +160,7 @@ public:
      * @param expectedNetworkOutput The desired network output.
      * @return Return true if operation was successful. Otherwise false
      */
-    bool computeBackprogationError(const Eigen::VectorXf& errorNextLayer, const Eigen::MatrixXf& weightMatrixNextLayer );
+    bool computeBackprogationError(const Eigen::VectorXd& errorNextLayer, const Eigen::MatrixXd& weightMatrixNextLayer );
 
     /**
      * Computes the partial derivatives of the biases and weights. Results can be
@@ -173,25 +173,25 @@ public:
      * derivatives and the learning rate.
      * @param eta Learning rate
      */
-    void updateWeightsAndBiases( const float& eta );
+    void updateWeightsAndBiases( const double& eta );
 
     /**
      * Returns the computed backprogation error in this layer.
      * @return
      */
-    const Eigen::VectorXf getBackpropagationError() const { return m_backpropagationError; } 
+    const Eigen::VectorXd getBackpropagationError() const { return m_backpropagationError; }
 
     /**
      * Partial derivatives of the biases. This is set after calling computePartialDerivatives();
      * @return
      */
-    const Eigen::VectorXf getPartialDerivativesBiases() const { return m_bias_partialDerivatives; }
+    const Eigen::VectorXd getPartialDerivativesBiases() const { return m_bias_partialDerivatives; }
 
     /**
      * Partial derivatives of weights. This is set after calling computePartialDerivatives();
      * @return
      */
-    const Eigen::MatrixXf getPartialDerivativesWeights() const { return m_weight_partialDerivatives; }
+    const Eigen::MatrixXd getPartialDerivativesWeights() const { return m_weight_partialDerivatives; }
 
     /**
      * Computes component wise derrivative of the sigmoid function
@@ -199,7 +199,7 @@ public:
      * @param z Vector in
      * @return Vector holding the result.
      */
-    static const Eigen::VectorXf d_sigmoid( const Eigen::VectorXf& z );
+    static const Eigen::VectorXd d_sigmoid( const Eigen::VectorXd& z );
 
     unsigned int getNbrOfNeurons() const { return m_nbr_of_neurons; }
     unsigned int getNbrOfNeuronInputs() const { return m_nbr_of_inputs; }
@@ -216,23 +216,23 @@ private:
      * @param activation_out
      * @return True if successful.
      */
-    bool setActivationOutput( const Eigen::VectorXf& activation_out );
+    bool setActivationOutput( const Eigen::VectorXd& activation_out );
 
 
 private:
     const unsigned int m_nbr_of_neurons;
     const unsigned int m_nbr_of_inputs;
 
-    Eigen::VectorXf m_activation_in;
-    Eigen::VectorXf m_activation_out;
-    Eigen::VectorXf m_z_weighted_input;
+    Eigen::VectorXd m_activation_in;
+    Eigen::VectorXd m_activation_out;
+    Eigen::VectorXd m_z_weighted_input;
 
-    Eigen::VectorXf m_backpropagationError;
-    Eigen::MatrixXf m_weightMatrix;
-    Eigen::VectorXf m_biasVector;
+    Eigen::VectorXd m_backpropagationError;
+    Eigen::MatrixXd m_weightMatrix;
+    Eigen::VectorXd m_biasVector;
 
-    Eigen::VectorXf m_bias_partialDerivatives;
-    Eigen::MatrixXf m_weight_partialDerivatives;
+    Eigen::VectorXd m_bias_partialDerivatives;
+    Eigen::MatrixXd m_weight_partialDerivatives;
 };
 
 #endif //LAYERHEADER
